% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_mess_all.R
\name{clean_mess_all}
\alias{clean_mess_all}
\title{Clean the Messenger Data Text and Time in One Go}
\usage{
clean_mess_all(data, custom_clean = NULL, timezone = "America/Los_Angeles")
}
\arguments{
\item{data}{The original messenger data that was selected using selection()}

\item{custom_clean}{An optional character vector which contains personal
stop words. For example, if you'd like to remove "http", it could be added.}

\item{timezone}{A character vector with the timezone of the user. Default
is Pacific Time.}
}
\value{
A tibble with cleaned up text and time data.
}
\description{
This function cleans the content of the text that was sent for each message.
Specifically, it removes special characters, turns all words to lowercase,
removes stopwords (words such as the, and, etc.), and removes white space.

A column is also created ("convo") which denotes who sent the message and
to which chat.

Messages that have text longer than 640 characters are removed, as
640 is the character limit for Messenger. Anything longer is bogus.

The cleaned version of the text is stored in the original content vector. The
original message content is also retained, but stored in the
vector: "content_unclean".

Additionally, it cleans up the time of the selected messenger data
so that it is more easily usable in R. It also adds columns for date and
hour for easy analysis.
}
