% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_mess_text.R
\name{clean_mess_text}
\alias{clean_mess_text}
\title{Clean the messenger data text content

This function cleans the content of the text that was sent for each message.
Specifically, it removes special characters, turns all words to lowercase,
removes stopwords (words such as the, and, etc.), and removes white space.

A column is also created ("convo") which denotes who sent the message and
to which chat.

Messages that have text longer than 640 characters are removed, as
640 is the character limit for Messenger. Anything longer is bogus.

The cleaned version of the text is stored in the original content vector. The
original message content is also retained, but stored in the
vector: "content_unclean".}
\usage{
clean_mess_text(data)
}
\arguments{
\item{data}{The original messenger data that was selected using selection()}
}
\value{
A data frame with cleaned up text data.
}
\description{
Clean the messenger data text content

This function cleans the content of the text that was sent for each message.
Specifically, it removes special characters, turns all words to lowercase,
removes stopwords (words such as the, and, etc.), and removes white space.

A column is also created ("convo") which denotes who sent the message and
to which chat.

Messages that have text longer than 640 characters are removed, as
640 is the character limit for Messenger. Anything longer is bogus.

The cleaned version of the text is stored in the original content vector. The
original message content is also retained, but stored in the
vector: "content_unclean".
}
