% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_mess_time.R
\name{clean_mess_time}
\alias{clean_mess_time}
\title{Clean selected Facebook Messenger data

This function cleans up the selected Facebook Messenger data after using the
selection() function.

This will convert the time-stamp to a usable format.}
\usage{
clean_mess_time(data)
}
\arguments{
\item{data}{The original messenger data that was selected using selection()}
}
\value{
A data frame with time stamp fixed to be usable in R, and date and
time columns.
}
\description{
Clean selected Facebook Messenger data

This function cleans up the selected Facebook Messenger data after using the
selection() function.

This will convert the time-stamp to a usable format.
}
