% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grady_augmented.R
\docType{data}
\name{grady_augmented}
\alias{grady_augmented}
\title{Grady-Augmented Dictionary Words}
\format{
A character vector with 122806 elements.
}
\source{
Taken from the Lexicon package
}
\usage{
data(grady_augmented)
}
\description{
Grady-Augmented Dictionary Words
}
\keyword{datasets}
