% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{group_sum}
\alias{group_sum}
\title{Group Chat Message Summary}
\usage{
group_sum(data)
}
\arguments{
\item{data}{A cleaned data frame of messenger data
\strong{Note - Must be performed before sentiment is added to the dataset}}
}
\value{
A tibble summarizing the total number of messages and the total
number of characters sent in a specific group chat in the dataset.
}
\description{
Group Chat Message Summary
}
