% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mess_sentiment.R
\name{mess_sentiment}
\alias{mess_sentiment}
\title{Find the Associated Sentiment of Each Word for Selected Messages}
\usage{
mess_sentiment(data, lexicon = "afinn")
}
\arguments{
\item{data}{Messenger data selected using the mess_selection() function and
cleaned using the clean_mess_text() or clean_mess_all() functions.}

\item{lexicon}{A character vector with that matches the desired input of
get_sentiments options c("bing", "afinn", "loughran", "nrc")). Default is
afinn. See the tidytext package for more details.}
}
\value{
A tibble in long format with words parsed and scored
}
\description{
This function adds sentiment scoring to each word of the
selected messenger data. This must be performed after running the mess_select
and clean_mess_text functions.
}
