% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_fun.R
\name{plot_mess}
\alias{plot_mess}
\title{Plot Sentiment Analysis Over Time per Sender}
\usage{
plot_mess(data)
}
\arguments{
\item{data}{The original messenger data that was selected using selection()}
}
\value{
A ggplot column-plot of the sentiment before and after an event,
and a density plot of each sender's overall sentiment before and after the
event date specified.
}
\description{
This function creates ggplots of Messenger data that have been
run through the get_mess_sentiment function.
}
