% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_fun.R
\name{plot_mess_dens}
\alias{plot_mess_dens}
\title{Plot Sentiment Analysis Over Time per Sender}
\usage{
plot_mess_dens(data, wrap_by = "sender")
}
\arguments{
\item{data}{The original messenger data that was selected using selection()}

\item{wrap_by}{The variable by which the ggplot will facet wrap}
}
\value{
A ggplot density plot of each sender's overall sentiment before and
after the event date specified.
}
\description{
This function creates ggplots of Messenger data that have been
run through the get_mess_sentiment function.
}
