% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_fun.R
\name{plot_senti_time}
\alias{plot_senti_time}
\title{Plot Sentiment Analysis Over Time}
\usage{
plot_senti_time(data)
}
\arguments{
\item{data}{The original messenger data that was selected using selection()}
}
\value{
A ggplot column-plot of the sentiment before and after an event,
and a density plot of each sender's overall sentiment before and after the
event date specified.
}
\description{
This function creates ggplots of Messenger data that have been
run through get_mess_sentiment() and through mess_cut_off(). Facet wrapping
is applied tomake multiple plots, with "sender" being the default variable
to wrap.
}
