% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{sender_sum}
\alias{sender_sum}
\title{Sender Summary}
\usage{
sender_sum(data)
}
\arguments{
\item{data}{A cleaned data frame of messenger data.
\strong{Note - Must be performed before sentiment is added to the dataset}}
}
\value{
A tibble summarizing the total number of messages and the total
number of characters sent by each person, to each specific chat
in the data set.
}
\description{
Sender Summary
}
