% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop_words.R
\docType{data}
\name{stop_words}
\alias{stop_words}
\title{Stop Words}
\format{
A data frame with 1149 rows and 2 variables:
word - An English word lexicon - The source of the stop word. Either "onix", "SMART", or "snowball".
}
\source{
Taken from the tidytext package
}
\usage{
data(stop_words)
}
\description{
Stop Words
}
\keyword{datasets}
