% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_fun.R
\name{t_mess}
\alias{t_mess}
\title{Run T Test on Messenger Data With Date Cut-Off Described.}
\usage{
t_mess(data)
}
\arguments{
\item{data}{Messenger data that has been cleaned using either the
clean_mess_text() and clean_mess_time() functions, or cleaned using the
clean_mess_all() function. Additionally, it must have been run through the
mess_cut_off() and the mess_sentiment() functions}
}
\value{
A tibble in long format filtered by top users
}
\description{
This function runs a t.test on messenger data that has been
cleaned and has been run through both mess_cut_off() and get_mess_sentiment()
It runs a t.test on the sentiment before and after the event time specified
in order to determine if there was a difference in word sentiment after the
event time.
}
