% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{top_words}
\alias{top_words}
\title{Top Words By User}
\usage{
top_words(data, num)
}
\arguments{
\item{data}{A cleaned data frame of messenger data
\strong{Note - Must be performed before sentiment is added to the dataset}}
}
\value{
A tibble of the number of dictionary words used by each person and
the total number of times used.
}
\description{
Top Words By User
}
