% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_off.R
\name{mess_cut_off}
\alias{mess_cut_off}
\title{Describe messages as "pre" or "post" an event}
\usage{
mess_cut_off(data, time)
}
\arguments{
\item{data}{The original messenger data that was selected using selection()
and cleaned using clean_mess_time() or clean_mess_all().}

\item{time}{A character vector with the selected date (for
example: "2000-12-31)}
}
\value{
A tibble in long format with messages described as "pre" or "post"
an event.
}
\description{
This function adds an event variable as "pre" or "post" a time set by user
to be used to measure for statistical tests on Facebook Messenger data after using the
selection() and clean_mess_time().
}
