% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mess_selection.R
\name{mess_selection}
\alias{mess_selection}
\title{Select FB Messenger Data

This function selects raw Facebook Messenger data. All that's needed are your
Facebook Messenger folders unzipped. The output is a data frame of
every message's content, sender and time stamp.}
\usage{
mess_selection(infile)
}
\arguments{
\item{infile}{The file path where the messenger data is stored. If NULL, will
use working directory.}
}
\value{
A data frame of each message's sender, content and time-stamp.
}
\description{
Select FB Messenger Data

This function selects raw Facebook Messenger data. All that's needed are your
Facebook Messenger folders unzipped. The output is a data frame of
every message's content, sender and time stamp.
}
